Teensy 3.2 INO code for SNES to USB conversion
My latest arduino (v.1.8.3) code file.  no restrictions.

I wrote most of the code myself, but this is really the compilation of a LOT of information from around the web.  It's also sort of an extenstion from the Instructables posting by johnburkert.  The primary difference is that john only included a hex file, while I am posting the arduino code.  As well, he used a teensy 2.0 and I used a 3.2 version.  different pinouts, naming conventions, etc.